#5 - Added capability to explicitly define default reminder list.
Alter the assignment of myDefaultList at the top of the script to anything
non-empty and the extension will now use the list of the given name as
default list. If nothing is re-configured the default list is still
selected automatically.

Integrated changes into the plain source as well as the alfred-reminders.scpt.